+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [2.2.0] - 2024-09-20
+
 ### Added
 
 - Configurable marker for highlighting the matched part of line
@@ -50,6 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release with core functionality
 
+[2.2.0]: https://github.com/Maarrk/silverbullet-grep/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/Maarrk/silverbullet-grep/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/Maarrk/silverbullet-grep/compare/v1.1.0...v2.0.0
 [1.1.0]: https://github.com/Maarrk/silverbullet-grep/compare/v1.0.0...v1.1.0